RDKB-43176: dropbear is not running, need restart logs seen in selfheal
Reason for change: Disable monitoring dropbear process

Test Procedure:
1. Verify that thee is not timeout when Device.Ethernet is queried

Risks: Low
(cherry picked from commit adbc9c5cea46b6357d67d83a123f03bfee6b4dc8)

+++ source/scripts/init/service.d/service_sshd.sh
@@ -241,8 +241,12 @@ service_start() {
 	    	do_start
 		fi
    fi
-		$PMON setproc ssh dropbear $PID_FILE "/etc/utopia/service.d/service_sshd.sh sshd-restart"
+   #Disable monitoring dropbear as we don't have a seperate dropbear process running always
+   #dropbear process would be running on demand basis
+   if [ "$BOX_TYPE" != "HUB4" ] && [ "$BOX_TYPE" != "SR300" ] && [ "$BOX_TYPE" != "SE501" ] && [ "$BOX_TYPE" != "WNXL11BWL" ] && [ "$BOX_TYPE" != "SR213" ]; then
 
+                $PMON setproc ssh dropbear $PID_FILE "/etc/utopia/service.d/service_sshd.sh sshd-restart"
+    fi
 		sysevent set ${SERVICE_NAME}-errinfo
 		sysevent set ${SERVICE_NAME}-status "started"
 		rm -rf /tmp/.dropbear/*
